+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `co-base-count`: Fix the feature not working after the gateway update
 - `cxpo-auto-price`: Fix localized number formatting
 - `hide-system-chat-messages`: Fix the vertical indicator not being visible after the gateway update
+- `other-context-notification-count`: Fix ghost INFRASTRUCTURE_UPGRADE_COMPLETED notifications
 
 ## 25.11.16